bug bytecode runner

+++ src/ByteCode.cpp
@@ -111,7 +111,7 @@ void ByteCode::enterByteCode(ByteCodeRunContext* context, uint32_t popParamsOnRe
 
     context->popParamsOnRet.push_back(popParamsOnRet);
     context->returnRegOnRet.push_back(returnReg);
-    if (returnReg != UINT8_MAX)
+    if (returnReg != UINT32_MAX)
         context->returnRegOnRetRR.push_back(context->registersRR[0].u64);
 }
